Changed generation of the projections and rwp
Directly appending a 0 or a 1 as a char to the bitmap instead of first creating a string of ints
Changed so the 1s are uniformly distributed
and random via the Mersenne Twister 19937 generator

set j to the numbers of ones and k to the negative number of zeros wanted
if the random number (that is between 0.0 and 1.0) is bigger than 0.5 and j is still not 0, append
getOneOrZero(j--) which gives getOneOrZero a positive number. The function then returns
 a char '1'. j gets decremented after being passed to the function. Also do this, in case k is 0,
 since this must mean, there are no possible number of 0s and thus 1s are needed now.
Do the same with random < 0.5, but give the negative k to getOneOrZero, which then returns
'0', while k is incremented.